Tablas de SRS, comentarios informativos
Se implementaron matrices que constituyen los datos para cada tetrominó respecto al Super Rotation System (SRS), que consiste en determinar cuánto se desfazan los centros de los bloques acorde a una rotación. Se harían varias pruebas de rotación recorriedo los array con los datos, y en caso de poder rotar se colocará la pieza.
