+++ src/Components/Tetros/Tetromino_O.java
@@ -2,23 +2,23 @@ package Components.Tetros;
 
 import GUI.Properties.TColor;
 
-public class Tetrominoe_Z extends Tetrominoe {
+public class Tetromino_O extends Tetromino {
     /*
         Shape (3 x 3)
         0º      90º     180º    270º
         r0      r1      r2      r3
-        ■ ■ □   □ □ ■   □ □ □   □ ■ □
-        □ ■ ■   □ ■ ■   ■ ■ □   ■ ■ □
-        □ □ □   □ ■ □   □ ■ ■   ■ □ □
+        □ ■ ■   □ ■ ■   □ ■ ■   □ ■ ■
+        □ ■ ■   □ ■ ■   □ ■ ■   □ ■ ■
+        □ □ □   □ □ □   □ □ □   □ □ □
     */
     private static final byte[][] spawnState = {
-            {7, 7, 0},
-            {0, 7, 7},
+            {0, 4, 4},
+            {0, 4, 4},
             {0, 0, 0},
     };
     private final byte[][] rotateShape = new byte[4][4];
-    public Tetrominoe_Z () {
-        tetroColor = TColor.RED;
+    public Tetromino_O() {
+        tetroColor = TColor.YELLOW;
     }
 
     @Override
