+++ src/Components/Tetros/Tetromino_T.java
@@ -2,23 +2,23 @@ package Components.Tetros;
 
 import GUI.Properties.TColor;
 
-public class Tetrominoe_O extends Tetrominoe {
+public class Tetromino_T extends Tetromino {
     /*
         Shape (3 x 3)
         0º      90º     180º    270º
         r0      r1      r2      r3
-        □ ■ ■   □ ■ ■   □ ■ ■   □ ■ ■
-        □ ■ ■   □ ■ ■   □ ■ ■   □ ■ ■
-        □ □ □   □ □ □   □ □ □   □ □ □
+        □ ■ □   ■ □ □   □ □ □   □ ■ □
+        ■ ■ ■   ■ ■ □   ■ ■ ■   ■ ■ □
+        □ □ □   ■ □ □   □ ■ □   □ ■ □
     */
     private static final byte[][] spawnState = {
-            {0, 4, 4},
-            {0, 4, 4},
+            {0, 6, 0},
+            {6, 6, 6},
             {0, 0, 0},
     };
     private final byte[][] rotateShape = new byte[4][4];
-    public Tetrominoe_O() {
-        tetroColor = TColor.YELLOW;
+    public Tetromino_T() {
+        tetroColor = TColor.VIOLET;
     }
 
     @Override
@@ -29,7 +29,7 @@ public class Tetrominoe_O extends Tetrominoe {
     public byte[][] getRotateShape() {
         return rotateShape;
     }
-    public byte[][] getSpawnState() {
+    public byte[][] getSpawnShape() {
         return spawnState;
     }
 }
